Fix bug with PE .raw.tif reads
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@3152 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ imports/G2img_1TIF.py
@@ -276,7 +276,7 @@ def GetTifData(filename):
 #                    image = np.fromfile(File,dtype=np.float32,count=4*Npix)[:Npix]
 #                    image = np.array(ar.array('f',File.read(4*Npix)),dtype=np.int32)
                 else:
-                    image = np.array(np.frombuffer(File.read(4*Npix),dtype=np.int),dtype=np.int32)
+                    image = np.array(np.frombuffer(File.read(4*Npix),dtype=np.int32),dtype=np.int32)
 #                    image = np.fromfile(File,dtype=np.int,count=4*Npix)[:Npix]
 #                    image = np.array(ar.array('I',File.read(4*Npix)),dtype=np.int32)
             elif IFD[258][2][0] == 16:
